polished code
