+++ includes/minishell.h
@@ -68,7 +68,6 @@ typedef struct s_tree_node
 {
 	t_node_type type;
 	char **argv;
-	//char *file_name;
 	struct s_tree_node *left;
 	struct s_tree_node *right;
 } t_tree_node;
@@ -92,7 +91,10 @@ int    		call_external_command(char **command, t_shell *shell);
 int			cd(const char *path, t_shell *shell);
 void		check_if_token_sequence(t_token **priora, t_token **priora_end, t_token_type cur_token_type);
 int     	check_operator(char *token);
+void		check_ptr(void *ptr);
 void		close_file(int fd);
+void		close_streams(t_shell *shell, char option);
+void		collect_heredocs(t_tree_node *node, t_shell *shell);
 pid_t		create_subprocess(void);
 t_tree_node *create_tree_node(t_token *token);
 void		delete_var(char *var_name, char ***var_store);
@@ -113,6 +115,8 @@ void		init_shell(t_shell *shell, char **envp);
 void		init_token_priority(t_priora *prior);
 int			is_symbol_oper(char с);
 t_token		*lexer(char *input);
+char		*ms_strdup(const char *str);
+char		*ms_strjoin(const char *s1, const char *s2);
 int			open_file(char *filename, int flags);
 void		overwrite_stream(char stream, int *info, int new_fd);
 t_tree_node *parser(char *input);
@@ -121,11 +125,13 @@ void		reset_streams(t_shell shell);
 int			run_cmd_in_current_process(int fd_to_duplicate, int fd[2], t_tree_node *node, t_shell *shell);
 char		*run_here_doc(char *stop_str_with_quotes, t_shell *shell);
 int			same_string(char *str1, char *str2);
-void		save_streams(int input, int output, t_attr *attributes, t_attr *attrib_orig_lnk);
+void		save_current_streams(t_shell *shell);
 int			search_var(char *var_name, char **var_store);
+void		set_streams(t_shell *shell, char option);
 void		set_var(char *var_name, char ***var_store);
 char		**shorten_arr(char *del_str, char **arr);
 void		tokenize_operator(char **str, t_token **token_list, t_priora priority_map);
+char		*trim_quotes(char *str, int *quoted);
 int			unset(char *var_name, t_shell *shell);
 void		wait_for_subprocess(pid_t pid);
