+++ srcs/redirections.c
@@ -51,21 +51,11 @@ static char *process_filename_node(t_tree_node *node)
 	return (filename);
 }
 
-// static void reset_heredoc_file(char *filename)
-// {
-// 	int fd;
-
-// 	fd = open(filename, O_TRUNC);
-// 	close(fd);
-// 	unlink(filename);
-// }
-
 int execute_redirection(t_tree_node *node, t_shell *shell, int streams)
 {
 	int status;
 	char direction;
 	char *filename;
-	//char *heredoc_file;
 
 	filename = process_filename_node(node->right);
 	direction = get_direction(node->type);
@@ -75,11 +65,6 @@ int execute_redirection(t_tree_node *node, t_shell *shell, int streams)
 		redirect_input(filename, &streams, 'H');
 	else if (direction == 'O')
 		redirect_output(node, filename, &streams);
-	// else if (direction == 'H')
-	// {
-	// 	heredoc_file = run_here_doc(filename, shell);
-	// 	redirect_input(heredoc_file, &streams);
-	// }
 	status = execute_command_line(node->left, shell, 0/*?*/, streams);
 	if (direction == 'H')
 		unlink(filename);
